+++ spring-cloud-tencent-examples/tsf-example/msgw-scg/src/main/java/com/tencent/cloud/tsf/msgw/scg/ScgApplication.java
@@ -1,0 +1,17 @@
+package com.tencent.cloud.tsf.msgw.scg;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.tsf.annotation.EnableTsf;
+
+/**
+ * @author seanlxliu
+ */
+@SpringBootApplication
+@EnableTsf
+public class ScgApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ScgApplication.class, args);
+	}
+}