+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/instrument/resttemplate/EnhancedRestTemplateWrapInterceptor.java
@@ -58,7 +58,7 @@ public class EnhancedRestTemplateWrapInterceptor {
 	}
 
 
-	public <T> T intercept(HttpRequest httpRequest, String serviceId, ServiceInstance serviceInstance,
+	public <T> T intercept(HttpRequest httpRequest, byte[] body, String serviceId, ServiceInstance serviceInstance,
 			LoadBalancerRequest<T> loadBalancerRequest) throws IOException {
 
 		EnhancedPluginContext enhancedPluginContext = new EnhancedPluginContext();
@@ -76,6 +76,9 @@ public class EnhancedRestTemplateWrapInterceptor {
 				.build();
 		enhancedPluginContext.setRequest(enhancedRequestContext);
 		enhancedPluginContext.setOriginRequest(httpRequest);
+		if (body != null) {
+			enhancedPluginContext.setOriginBody(body);
+		}
 
 		enhancedPluginContext.setLocalServiceInstance(pluginRunner.getLocalServiceInstance());
 
@@ -104,7 +107,7 @@ public class EnhancedRestTemplateWrapInterceptor {
 				enhancedResponseContextBuilder.httpStatus(((ClientHttpResponse) response).getStatusCode().value());
 				enhancedResponseContextBuilder.httpHeaders(((ClientHttpResponse) response).getHeaders());
 			}
-			EnhancedResponseContext enhancedResponseContext =  enhancedResponseContextBuilder.build();
+			EnhancedResponseContext enhancedResponseContext = enhancedResponseContextBuilder.build();
 
 			enhancedPluginContext.setResponse(enhancedResponseContext);