+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [feat:support setting load balancing strategy per service.](https://github.com/Tencent/spring-cloud-tencent/pull/1633)
 - [feat: support tsf 2024.](https://github.com/Tencent/spring-cloud-tencent/pull/1635)
 - [feat: support shortest response time lb and least connection lb](https://github.com/Tencent/spring-cloud-tencent/pull/1637)
+- [feat:support traffic mirroring.](https://github.com/Tencent/spring-cloud-tencent/pull/1647)